+++ PROJECT
@@ -36,6 +36,6 @@ resources:
   domain: kyma-project.io
   group: gateway
   kind: RateLimit
-  path: github.com/kyma-project/api-gateway/apis/gateway/v1alpha1
+  path: github.com/kyma-project/api-gateway/apis/gateway/ratelimit/v1alpha1
   version: v1alpha1
 version: "3"
