+++ controllers/gateway/ratelimit/ratelimit_enabled.go
@@ -1,9 +1,9 @@
 //go:build dev_features
 
-package gateway
+package ratelimit
 
 import (
-	gatewayv1alpha1 "github.com/kyma-project/api-gateway/apis/gateway/v1alpha1"
+	gatewayv1alpha1 "github.com/kyma-project/api-gateway/apis/gateway/ratelimit/v1alpha1"
 	"k8s.io/apimachinery/pkg/runtime"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
